Adding in bible-plan.core seems to have screwed up the auto-connection of the brepl again. Whatevs. I can now use goog.date.DateTime with abandon!
Remember that adding a new require is going to require recompiling the source as the library has to be included in the browser in order to be useful to my code.

Not quite what I'm used to in JVM land but still not as bad as it could be. Think of adding a new require like you're adding a new dependency.